+++ CHANGES.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Rename `volume` setting to `master volume`
 * Tweak island 62 to make it possible for leaf to win
 * Display more info in the map editor
+* Tune AI
+  * It will try to attack enemy living pieces in the order of baron, knight, spearman, peasant
 
 ### Fixed